pragmatic/settings/local.py
from .base import *

import environ

environ.Env.read_env(os.path.join(BASE_DIR, '.env'))

env = environ.Env(
    DEBUG=(bool, False)
)

# Quick-start development settings - unsuitable for production
# See https://docs.djangoproject.com/en/4.0/howto/deployment/checklist/

# SECURITY WARNING: keep the secret key used in production secret!
SECRET_KEY = env('SECRET_KEY')

# SECURITY WARNING: don't run with debug turned on in production!
DEBUG = True

ALLOWED_HOSTS = []


# Database
# https://docs.djangoproject.com/en/4.0/ref/settings/#databases

import pymysql

pymysql.install_as_MySQLdb()

DATABASES = {
    'default': {
        'ENGINE': 'django.db.backends.sqlite3',
        'NAME': BASE_DIR / 'db.sqlite3',
    },
    'mysqldb': {
        'ENGINE': 'django.db.backends.mysql',
        'NAME': 'pragmatic',
        'USER': 'kwon',
        # 'PASSWORD' : 'kwon0711',
        'PASSWORD': 'charlie',
        'HOST': '127.0.0.2',
        'PORT': 3306,
    }
}
