requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
Django==4.0.6
django-bootstrap4==22.1
django-environ==0.9.0
mysqlclient==2.1.1
Pillow==9.2.0
PyMySQL==1.0.2
PyMysqlDB==0.0.2
soupsieve==2.3.2.post1
sqlparse==0.4.2
